requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
dependency-injector==4.40.0
et-xmlfile==1.1.0
fixture==1.5.11
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
h11==0.14.0
httpcore==0.16.1
httplib2==0.21.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
install==1.3.5
loguru==0.6.0
multidict==6.0.3
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pluggy==1.0.0
prettytable==3.5.0
protobuf==4.21.11
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-asyncio==0.20.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.0a6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
sentry-sdk==1.11.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
yarl==1.8.2
